KIWI-1744: Athena Test Harness + QA Logic
