+++ src/tests/api/BavUnhappyPath.test.ts
@@ -10,13 +10,10 @@ import {
 	verifyAccountPost,
 	tokenPost,
 	getSessionAndVerifyKey,
-	getAthenaRecordByFirstNameAndTime,
 } from "./ApiTestSteps";
 import { constants } from "./ApiConstants";
 import { BankDetailsPayload } from "../models/BankDetailsPayload";
 import { randomUUID } from "crypto";
-import exampleStubPayload from "../data/exampleStubPayload.json";
-import { absoluteTimeNow, sleep } from "./ApiUtils";
 
 describe("BAV CRI unhappy path tests", () => {
 	describe("/session Endpoint Unhappy Path Tests", () => {
@@ -103,29 +100,6 @@ describe("BAV CRI unhappy path tests", () => {
 
 			expect(verifyAccountResponseSecondRetry.status).toBe(401);
 			expect(verifyAccountResponseSecondRetry.data).toBe("Too many attempts");
-		});
-	});
-
-	describe("BAV CRI: /verify-account Endpoint Partial Match Athena Output Test", () => {
-		it("Triggers Partial Match and checks for object in Athena", async () => {
-			// ARRANGE
-			const firstName = randomUUID().slice(-8);
-			const newStubPayload = structuredClone(exampleStubPayload);
-			newStubPayload.shared_claims.name[0].nameParts[0].value = firstName;
-
-			const sessionId = await startStubServiceAndReturnSessionId(newStubPayload);
-
-			const newVerifyAccountYesPayload = structuredClone(verifyAccountYesPayload);
-			newVerifyAccountYesPayload.account_number = "00111114";
-			const startTime = absoluteTimeNow();
-
-			// ACT
-			const verifyAccountResponse = await verifyAccountPost(newVerifyAccountYesPayload, sessionId);
-
-			// ASSERT
-			expect(verifyAccountResponse.status).toBe(200);
-			const athenaRecords = await getAthenaRecordByFirstNameAndTime(startTime, firstName);
-			expect(athenaRecords.length).toBeGreaterThan(0);
 		});
 	});
